+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Check for metric file in addition to item directory in CheckpointManager.
 - Additional logs to indicate save/restore completion.
 - Support for None leaves in PyTree save/restore.
 - `read: bool` option on all_steps to force read from storage location instead